MarkovKernels

Documentation for MarkovKernels.

MarkovKernels.AbstractNormal

MarkovKernels.AbstractNormalKernel

MarkovKernels.Normal

MarkovKernels.NormalKernel

MarkovKernels.condition

MarkovKernels.covp

MarkovKernels.covp

MarkovKernels.dim

MarkovKernels.entropy

MarkovKernels.kldivergence

MarkovKernels.logpdf

MarkovKernels.residual

MarkovKernels.AbstractNormal — Type

AbstractNormal{T<:Number}

Abstract type for representing normal distributed random vectors taking values in T.

MarkovKernels.Normal — Type

Normal{T,U,V}

Standard parametrisation of the normal distribution with element type T.

MarkovKernels.covp — Method

covp(N::Normal)

Returns the internal representation of the covariance matrix of the Normal distribution N. For computing the actual covariance matrix, use cov.

MarkovKernels.dim — Method

dim(N::AbstractNormal)

Returns the dimension of the normal distribution N.

MarkovKernels.entropy — Method

entropy(N::AbstractNormal)

Returns the entropy of N.

MarkovKernels.kldivergence — Method

kldivergence(N1::AbstractNormal,N2::AbstractNormal)

Returns the Kullback-Leibler divergence between N1 and N2.

MarkovKernels.logpdf — Method

logpdf(N::AbstractNormal,x)

Returns the logarithm of the probability density function of N evaluated at x.

MarkovKernels.residual — Method

residual(N::AbstractNormal,x)

Returns the whitened residual associated with N and observed vector x.

MarkovKernels.AbstractNormalKernel — Type

AbstractNormalKernel{T<:Number}

Abstract type for representing Normal kernels taking values in T.

MarkovKernels.NormalKernel — Type

NormalKernel

Standard parametrisation of Normal kernels.

MarkovKernels.condition — Method

condition(K::AbstractNormalKernel, x)

Returns a Normal distribution corresponding to K evaluated at x.

MarkovKernels.covp — Method

covp(K::NormalKernel)

Returns the internal representation of the conditonal covariance matrix of the Normal kernel K. For computing the actual conditional covariance matrix, use cov.