Internal methods for covariance parameters

MarkovKernels.utrisqrt2utrichol! — Function

utrisqrt2utrichol!(A::AbstractMatrix)

Applies an in-place orthogonal transform to the upper triangular matrix A, such that the diagonal entries are real and positive, i.e. the resulting matrix is a valid Cholesky factor.

MarkovKernels.positive_qrwoq! — Method

positive_qrwoq!(A::AbstractMatrix)

Computes the R factor in the QR decomposition of A, in-place, ensuring that the diagonal entries of R are positive. The returned object is a view of A.