SelfAdjoint

Self-adjoint matrices are subtypes of the following union:

const RSym{T,S} = Symmetric{T,S} where {T<:Real,S}
const CHerm{T,S} = Hermitian{T,S} where {T<:Complex,S}
const SelfAdjoint{T,S} = Union{RSym{T,S},CHerm{T,S}} where {T,S}

Methods

MarkovKernels.selfadjoint! — Method

selfadjoint!(A)

Computes the self-adjoint part of A, in-place, and wraps it in an appropriate self-adjoitn wrapper type (i.e. Symemtric / Hermitian).

MarkovKernels.selfadjoint — Method

selfadjoint(A)

Computes the self-adjoint part of A and wraps it in an appropriate self-adjoitn wrapper type (i.e. Symemtric / Hermitian).

MarkovKernels.selfadjoint — Method

selfadjoint(A)

Computes the self-adjoint part of A and wraps it in an appropriate self-adjoitn wrapper type (i.e. Symemtric / Hermitian).

MarkovKernels.selfadjoint — Method

selfadjoint(A)

Computes the self-adjoint part of A and wraps it in an appropriate self-adjoitn wrapper type (i.e. Symemtric / Hermitian).

MarkovKernels.rsqrt — Method

rsqrt(A::SelfAdjoint)

Computes the rigtht square-root of A.

MarkovKernels.lsqrt — Method

lsqrt(A::SelfAdjoint)

Computes the left square-root of A.

MarkovKernels.stein — Method

stein(Σ::SelfAdjoint, Φ::AbstractMatrix)

Computes the output of the stein operator

Σ ↦ Φ * Σ * Φ'.

MarkovKernels.stein — Method

stein(Σ::SelfAdjoint, Φ::AbstractMatrix, Q::SelfAdjoint)

Computes the output of the stein operator

Σ ↦ Φ * Σ * Φ' + Q.

Both Σ and Q need to be of the same CovarianceParameter type, e.g. both SymOrHerm or both Cholesky. The type of the CovarianceParameter is preserved at the output.

MarkovKernels.stein — Method

stein(Σ::SelfAdjoint, Φ::AbstractMatrix, Q::SelfAdjoint)

Computes the output of the stein operator

Σ ↦ Φ * Σ * Φ' + Q.

Both Σ and Q need to be of the same CovarianceParameter type, e.g. both SymOrHerm or both Cholesky. The type of the CovarianceParameter is preserved at the output.

MarkovKernels.schur_reduce — Method

schur_reduce(Π::SelfAdjoint, C::AbstractMatrix)

Returns the tuple (S, K, Σ) associated with the following (block) Schur reduction:

[C*Π*C' C*Π; Π*C' Π] = [0 0; 0 Σ] + [I; K]*(C*Π*C')*[I; K]'

In terms of Kalman filtering, Π is the predictive covariance, C the measurement matrix, and R the measurement covariance, then S is the marginal measurement covariance, K is the Kalman gain, and Σ is the filtering covariance.

MarkovKernels.schur_reduce — Method

schur_reduce(Π::SelfAdjoint, C::AbstractMatrix, R::SelfAdjoint)

Returns the tuple (S, K, Σ) associated with the following (block) Schur reduction:

[C*Π*C'+R C*Π; Π*C' Π] = [0 0; 0 Σ] + [I; K]*(C*Π*C' + R)*[I; K]'

In terms of Kalman filtering, Π is the predictive covariance, C the measurement matrix, and R the measurement covariance, then S is the marginal measurement covariance, K is the Kalman gain, and Σ is the filtering covariance.

MarkovKernels.schur_reduce — Method

schur_reduce(Π::SelfAdjoint, C::AbstractMatrix, R::Real)

Returns the tuple (S, K, Σ) associated with the following (block) Schur reduction:

[C*Π*C'+R C*Π; Π*C' Π] = [0 0; 0 Σ] + [I; K]*(C*Π*C' + R)*[I; K]'

In terms of Kalman filtering, Π is the predictive covariance, C the measurement matrix, and R the measurement covariance, then S is the marginal measurement covariance, K is the Kalman gain, and Σ is the filtering covariance.