Affine maps

An affine map is a function $f$ given by

\[ f(x) = A x + b,\]

where $A$ is the slope and $b$ is the intercept. Different representations of affine maps are sometimes useful, as documented below.

Types

MarkovKernels.AbstractAffineMap — Type

AbstractAffineMap{T<:Number}

Abstract type for representing affine maps between vector spaces over the field determined by T.

MarkovKernels.AffineMap — Type

AffineMap{T,U,V}

Type for representing affine maps in the standard slope / intercept parametrisation.

MarkovKernels.LinearMap — Type

LinearMap{T,U}

Type for representing affine maps with zero intercept.

MarkovKernels.AffineCorrector — Type

AffineCorrector{T,U,V,S}

Type for representing affine correctors, i.e.,

x ↦ b + A * (x -c).

Constructors

MarkovKernels.AffineMap — Method

AffineMap(A::AbstractMatrix, b::AbstractVector)

Creates an AffineMap with slope A and intercept b.

MarkovKernels.LinearMap — Method

LinearMap(A::AbstractMatrix)

Creates a LinearMap with slope A.

MarkovKernels.AffineCorrector — Method

AffineCorrector(A::AbstractMatrix, b::AbstractVector, c::AbstractVector)

Creates an Affine Corrector with slope A and intercept b - A * c.

Basics

MarkovKernels.slope — Method

slope(F::AbstractAffineMap)

Computes the slope of F.

MarkovKernels.intercept — Method

intercept(F::AffineMap)

Computes the intercept of F.

MarkovKernels.compose — Method

compose(F2::AbstractAffineMap, F1::AbstractAffineMap)

Computes the affine map F3 resulting from the composition F2 ∘ F1.

Base.:* — Method

*(F2::AbstractAffineMap, F1::AbstractAffineMap)

Equivalent to compose(F2::AbstractAffineMap, F1::AbstractAffineMap).