Covariance Parameter

In MarkovKernels.jl a covariance matrix is assumed to be of the following type union.

const CovarianceParameter{T} = Union{HermOrSym{T},UniformScaling{T},Factorization{T}}

Functions

MarkovKernels.lsqrt — Method

lsqrt(A::CovarianceParameter)

Computes a square 'matrix' L such that A = L*L'. L need not be a Cholesky factor.

MarkovKernels.lsqrt — Method

lsqrt(A::AbstractMatrix)

Equivalent to cholesky(A).L

MarkovKernels.stein — Method

stein(Σ::CovarianceParameter, Φ::AbstractMatrix)

Computes the output of the stein operator

Σ ↦ Φ * Σ * Φ'.

Generally, the type of Σ determines the type of the output. The exception is the case when Q is a Factorization and Σ is of type Union{HermOrSym{T,Diagonal{T,V}},UniformScaling{T}}, in which case Q determines the type of the output.

MarkovKernels.stein — Method

stein(Σ::CovarianceParameter, Φ::AbstractMatrix, Q::CovarianceParameter)

Computes the output of the stein operator

Σ ↦ Φ * Σ * Φ' + Q.

Generally, the type of Σ determines the type of the output. The exception is the case when Q is a Factorization and Σ is of type Union{HermOrSym{T,Diagonal{T,V}},UniformScaling{T}}, in which case Q determines the type of the output.

MarkovKernels.schur_reduce — Method

schur_reduce(Π::CovarianceParameter, C::AbstractMatrix)

Returns the tuple (S, K, Σ) associated with the following (block) Schur reduction:

[C*Π*C' C*Π; Π*C' Π] = [0 0; 0 Σ] + [I; K]*(C*Π*C')*[I; K]'

In terms of Kalman filtering, Π is the predictive covariance, C the measurement matrix, and R the measurement covariance, then S is the marginal measurement covariance, K is the Kalman gain, and Σ is the filtering covariance.

Generally, the type of Π determines the type of the output. The exception is the case when R is a Factorization and Π is of type Union{HermOrSym{T,Diagonal{T,V}},UniformScaling{T}}, in which case R determines the type of the output.

MarkovKernels.schur_reduce — Method

schur_reduce(Π::CovarianceParameter, C::AbstractMatrix, R::CovarianceParameter)

Returns the tuple (S, K, Σ) associated with the following (block) Schur reduction:

[CΠC' + R CΠ; ΠC' Π] = [0 0; 0 Σ] + [I; K](CΠC' + R)[I; K]'

In terms of Kalman filtering, Π is the predictive covariance, C the measurement matrix, and R the measurement covariance, then S is the marginal measurement covariance, K is the Kalman gain, and Σ is the filtering covariance.

Generally, the type of Π determines the type of the output. The exception is the case when R is a Factorization and Π is of type Union{HermOrSym{T,Diagonal{T,V}},UniformScaling{T}}, in which case R determines the type of the output.