Cholesky

Methods

MarkovKernels.rsqrt — Method

rsqrt(C::Cholesky)

Computes the right-square root of C.

MarkovKernels.lsqrt — Method

lsqrt(C::Cholesky)

Computes the left-square root of C.

MarkovKernels.stein — Method

stein(Σ::Cholesky, Φ::AbstractMatrix)

Computes the output of the stein operator

Σ ↦ Φ * Σ * Φ'.

MarkovKernels.stein — Method

stein(Σ::Cholesky, Φ::Adjoint{<:Number,<:AbstractVector})

Computes the output of the stein operator

Σ ↦ Φ * Σ * Φ'.

The return type is a subtype of Real.

MarkovKernels.stein — Method

stein(Σ::Cholesky, Φ::AbstractMatrix, Q::Cholesky)

Computes the output of the stein operator

Σ ↦ Φ * Σ * Φ' + Q.

MarkovKernels.stein — Method

stein(Σ::Cholesky, Φ::Adjoint{<:Number,<:AbstractVector}, Q::Number)

Computes the output of the stein operator

Σ ↦ Φ * Σ * Φ' + Q.

The return type is a subtype of Real.

MarkovKernels.schur_reduce — Method

schur_reduce(Π::Cholesky, C::AbstractMatrix)

Returns the tuple (S, K, Σ) associated with the following (block) Schur reduction:

[C*Π*C' C*Π; Π*C' Π] = [0 0; 0 Σ] + [I; K]*(C*Π*C')*[I; K]'

In terms of Kalman filtering, Π is the predictive covariance, C the measurement matrix, and R the measurement covariance, then S is the marginal measurement covariance, K is the Kalman gain, and Σ is the filtering covariance.

MarkovKernels.schur_reduce — Method

schur_reduce(Π::Cholesky, C::Adjoint{<:Number,<:AbstractVector})

Returns the tuple (S, K, Σ) associated with the following (block) Schur reduction:

[C*Π*C' C*Π; Π*C' Π] = [0 0; 0 Σ] + [I; K]*(C*Π*C')*[I; K]'

In terms of Kalman filtering, Π is the predictive covariance, C the measurement matrix, and R the measurement covariance, then S is the marginal measurement covariance, K is the Kalman gain, and Σ is the filtering covariance.

MarkovKernels.schur_reduce — Method

schur_reduce(Π::Cholesky, C::AbstractMatrix, R::Cholesky)

Returns the tuple (S, K, Σ) associated with the following (block) Schur reduction:

[C*Π*C'+R C*Π; Π*C' Π] = [0 0; 0 Σ] + [I; K]*(C*Π*C' + R)*[I; K]'

In terms of Kalman filtering, Π is the predictive covariance, C the measurement matrix, and R the measurement covariance, then S is the marginal measurement covariance, K is the Kalman gain, and Σ is the filtering covariance.

MarkovKernels.schur_reduce — Method

schur_reduce(Π::Cholesky, C::Adjoint{<:Number,<:AbstractVector}, R::Number)

Returns the tuple (S, K, Σ) associated with the following (block) Schur reduction:

[C*Π*C'+R C*Π; Π*C' Π] = [0 0; 0 Σ] + [I; K]*(C*Π*C' + R)*[I; K]'

In terms of Kalman filtering, Π is the predictive covariance, C the measurement matrix, and R the measurement covariance, then S is the marginal measurement covariance, K is the Kalman gain, and Σ is the filtering covariance.