Dirac

The Dirac distribution with parameter $\mu$ is a distribution putting all probabiltiy mass on $\mu$. It is denoted by

\[\delta(x -\mu).\]

Types

MarkovKernels.AbstractDirac — Type

AbstractDirac{T<:Number}

Abstract type for representing Dirac random vectors taking values in T.

MarkovKernels.Dirac — Type

Dirac{T<:Number}

Type for representing Dirac random vectors taking values in T.

Constructors

MarkovKernels.Dirac — Method

Dirac(μ::AbstractVector)

Creates a Dirac distribution with mean vector μ.

MarkovKernels.Dirac — Method

Dirac{T}(D::Dirac{U,V})

Computes a Dirac distribution of eltype T from the Dirac distribution D if T and U are compatible. That is T and U must both be Real or both be Complex.

Basics

MarkovKernels.dim — Method

dim(D::AbstractDirac)

Returns the dimension of the Dirac distribution D.

Statistics.mean — Method

mean(D::AbstractDirac)

Computes the mean vector of the Dirac distribution D.

Missing docstring.

Missing docstring for cov(::AbstractDirac). Check Documenter's build log for details.

Missing docstring.

Missing docstring for var(::AbstractDirac). Check Documenter's build log for details.

Missing docstring.

Missing docstring for std(::AbstractDirac). Check Documenter's build log for details.

Sampling

Base.rand — Method

rand(RNG::AbstractRNG, D::AbstractDirac)

Computes a random vector distributed according to the Dirac distribution D using the random number generator RNG. Equivalent to mean(D).

Base.rand — Method

rand(D::AbstractDirac)

Computes a random vector distributed according to the Dirac distribution D using the random number generator Random.GLOBAL_RNG. Equivalent to mean(D).