Categorical

Types

MarkovKernels.AbstractCategorical — Type

AbstractCategorical{T}

Abstract type for representing categorical distributions with values T.

MarkovKernels.Categorical — Type

Categorical{T,A}

Type for representing categorical distributions with sample_eltype T.

Constructor

MarkovKernels.Categorical — Method

Categorical(p::AbstractVector)

Constructs a categorical distribution from the vector of probabilities p.

Basics

MarkovKernels.probability_vector — Method

probability_vector(::AbstractCategorical)

Computes the vector of probabilities for each category.