Dirac

The Dirac distribution with parameter $\mu$ is a distribution putting all probabiltiy mass on $\mu$. It is denoted by

\[\delta(x -\mu).\]

Types

MarkovKernels.AbstractDirac — Type

AbstractDirac{ST}

Abstract type for representing Dirac distributions taking values in ST.

MarkovKernels.Dirac — Type

Dirac

Type for representing Dirac distributions with sample_eltype ST.

Constructors

MarkovKernels.Dirac — Method

Dirac(μ)

Creates a Dirac distribution with mean μ.

Methods

MarkovKernels.dim — Method

dim(D::AbstractDirac)

Returns the dimension of the Dirac distribution D.

Statistics.mean — Method

mean(D::AbstractDirac)

Computes the mean vector of the Dirac distribution D.