General

Type

MarkovKernels.AbstractDistribution — Type

AbstractDistribution{ST}

Abstract type for representing distributions with samples of type ST.

Type information

MarkovKernels.sample_type — Method

sample_type(D::AbstractDistribution)

Computes the type of samples from D, e.g. same as typeof(rand(D)).

MarkovKernels.sample_eltype — Method

sample_eltype(D::AbstractDistribution)

Computes the eltype of samples from D, e.g. same as eltype(rand(D)).

Probability densities

MarkovKernels.logpdf — Method

logpdf(D::AbstractDistribution, x)

Computes the logarithm of the probabilidty density of D, evaluated at x.

Sampling

Base.rand — Method

rand([rng], D::AbstractDistribution)

Draws one sample from D.