ParticleSystem

A Particle system is a mixture of Dirac distributions.

\[P(x) = \sum_{i= 1}^n w_i \delta(x - μ^{(i)})\]

It can also be used to represent a mixture of trajectories.

\[P(x) = \sum_{i= 1}^n w_i \delta(x_{1:T} - μ^{(i)}_{1:T})\]

Types

MarkovKernels.AbstractParticleSystem — Type

AbstractParticleSystem{T} <: AbstractDistribution{T}

abstract type for representing systems of particles.

MarkovKernels.ParticleSystem — Type

ParticleSystem{T,A,B} <: AbstractParticleSystem{T}

type for representing a standard particle system.

Constructors

MarkovKernels.ParticleSystem — Method

ParticleSystem(logws::AbstractVector{<:Real}, X::AbstractArray{<:AbstractVector{T}})

Creates a ParticleSystem with logarithm of the mixture weights given by logws and location parameters in X. If X is an AbstractVector the resulting object represents a classical Dirac mixtrue. Whereas if X is an AbstractMatrix, the resulting object represents a Dirac mixture over trajectories, where the row dimension represents time and the column dimension enumerates the particles.

Basics

MarkovKernels.dim — Method

dim(P::AbstractParticleSystem)

Returns the dimension of the particle system distribution P.

MarkovKernels.logweights — Method

logweights(P::AbstractParticleSystem)

Returns the logarithms of the mixture weights of the particle system P.

MarkovKernels.weights — Method

weights(P::AbstractParticleSystem)

Returns the mixture weights of the particle system P.

MarkovKernels.nparticles — Method

nparticles(P::AbstractParticleSystem)

Computes the number of particles in the particle system P.

MarkovKernels.particles — Method

particles(P::AbstractParticleSystem)

Returns the particle locations of the particle system P.

Statistics.mean — Method

mean(P::AbstractParticleSystem)

Computes the mean of the particle system distribution P.