Probability vector

Types

MarkovKernels.AbstractProbabilityVector — Type

AbstractProbabilityVector{ST}

Abstract type for representing categorical distributions with values ST.

MarkovKernels.ProbabilityVector — Type

ProbabilityVector{T,A}

Type for representing categorical distributions with sample_eltype T.

Constructor

MarkovKernels.ProbabilityVector — Method

ProbabilityVector(p::AbstractVector)

Constructs a categorical distribution from the vector of probabilities p.

Methods

MarkovKernels.probability_vector — Method

probability_vector(::AbstractProbabilityVector)

Computes the vector of probabilities for each category.

MarkovKernels.entropy — Method

entropy(C::AbstractProbabilityVector)

Computes the entropy of the categorical distribution C.

MarkovKernels.kldivergence — Method

kldivergence(C1::AbstractProbabilityVector, C2::AbstractProbabilityVector)

Computes the Kullback-Leibler divergence between the categorical distributions C1 and C2.