Likelihood

Types

MarkovKernels.FlatLikelihood — Type

FlatLikelihood

Type for representing flat likelihoods.

MarkovKernels.Likelihood — Type

Likelihood{U,V}

Type for representing a Likelihood associated with a kernel K(y, x) and a measurement y.

Constructors

MarkovKernels.Likelihood — Method

Likelihood(K::AbstractMarkovKernel, y)

Creates a Likelihood with measurement kernel K and measurement y.

Basics

MarkovKernels.measurement_model — Method

measurement_model(L::Likelihood)

Computes the measurement kernel K.

MarkovKernels.measurement — Method

measurement(L::Likelihood)

Computes the measurement y

Base.log — Method

log(L::Likelihood, x)

Computes the logarithm of the likelihood L at x.