Normal

The standard parametrisation of the Normal distribution is given by

\[\mathcal{N}(x ; \mu , \Sigma ),\]

where $\mu$ is the mean vector and $\Sigma$ is the covariance matrix. The exact expression for the probabiltiy density function depends on whether $x$ is vector with real or complex values, both are supported. For real valued vectors the density function is given by

\[\mathcal{N}(x ; \mu , \Sigma ) = |2\pi \Sigma|^{-1/2} \exp \Big( -\frac{1}{2}(x-\mu)^* \Sigma^{-1} (x-\mu) \Big),\]

whereas for complex valued vectors the density function is given by

\[\mathcal{N}(x ; \mu , \Sigma ) = |\pi \Sigma|^{-1} \exp \Big( -(x-\mu)^* \Sigma^{-1} (x-\mu) \Big).\]

Types

MarkovKernels.AbstractNormal — Type

AbstractNormal{T<:Number}

Abstract type for representing Normal distributed random vectors taking values in T.

MarkovKernels.Normal — Type

Normal{T,U,V}

Standard mean vector / covariance matrix parametrisation of the normal distribution with element type T.

Type aliases

const IsoNormal{T,U} = Normal{T,U,<:UniformScaling}

Constructors

MarkovKernels.Normal — Method

Normal(μ::AbstractVector, Σ::CovarianceParameter)

Creates a Normal distribution with mean vector μ and covariance matrix parametrised by Σ.

MarkovKernels.Normal — Method

Normal(μ::AbstractVector, Σ::AbstractMatrix)

Creates a Normal distribution with mean vector μ and covariance matrix Σ if Σ is Symmetric / Hermitian. Throws domain error otherwise.

MarkovKernels.IsoNormal — Method

IsoNormal(μ::AbstractVector, λ::Real)

Short-hand for Normal(μ, λ * I).

MarkovKernels.Normal — Method

Normal{T}(N::Normal{U,V,W})

Computes a Normal distribution of eltype T from the Normal distribution N if T and U are compatible. That is T and U must both be Real or both be Complex.

Basics

MarkovKernels.dim — Method

dim(N::AbstractNormal)

Returns the dimension of the Normal distribution N.

Statistics.mean — Method

mean(N::AbstractNormal)

Computes the mean vector of the Normal distribution N.

Statistics.cov — Method

cov(N::AbstractNormal)

Computes the covariance matrix of the Normal distribution N.

MarkovKernels.covp — Method

covp(N::AbstractNormal)

Returns the internal representation of the covariance matrix of the Normal distribution N. For computing the actual covariance matrix, use cov.

Statistics.var — Method

var(N::AbstractNormal)

Computes the vector of marginal variances of the Normal distribution N.

Statistics.std — Method

std(N::AbstractNormal)

Computes the vector of marginal standard deviations of the Normal distribution N.

Probability density function

MarkovKernels.residual — Method

residual(N::AbstractNormal, x::AbstractVector)

Computes the whitened residual associated with the Normal distribution N and observed vector x.

MarkovKernels.logpdf — Method

logpdf(N::AbstractNormal,x)

Computes the logarithm of the probability density function of the Normal distribution N evaluated at x.

Information theory

MarkovKernels.entropy — Method

entropy(N::AbstractNormal)

Computes the entropy of the Normal distribution N.

MarkovKernels.kldivergence — Method

kldivergence(N1::AbstractNormal, N2::AbstractNormal)

Computes the Kullback-Leibler divergence between the Normal distributions N1 and N2.

Sampling

Base.rand — Method

rand(RNG::AbstractRNG, N::AbstractNormal)

Computes a random vector distributed according to the Normal distribution N using the random number generator RNG.

Base.rand — Method

rand(N::AbstractNormal)

Computes a random vector distributed according to the Normal distribution N using the random number generator Random.GLOBAL_RNG.