MarkovKernels

Documentation for MarkovKernels.

MarkovKernels.AbstractNormal

MarkovKernels.AbstractNormalKernel

MarkovKernels.Normal

MarkovKernels.NormalKernel

Base.rand

Base.rand

MarkovKernels.condition

MarkovKernels.dim

MarkovKernels.entropy

MarkovKernels.kldivergence

MarkovKernels.logpdf

MarkovKernels.residual

Statistics.cov

Statistics.cov

Statistics.mean

Statistics.mean

Statistics.std

Statistics.var

MarkovKernels.AbstractNormal — Type

AbstractNormal{T<:Number}

Abstract type for representing normal distributed random vectors taking values in T.

MarkovKernels.Normal — Type

Normal{T,U,V}

Standard parametrisation of the normal distribution with element type T.

Base.rand — Method

rand(N::AbstractNormal)

Draws one random vector using the random number generator GLOBAL_RNG.

Base.rand — Method

rand(RNG::AbstractRNG, N::AbstractNormal)

Draws one random vector from N using the random number generator RNG.

MarkovKernels.dim — Method

dim(N::AbstractNormal)

Returns the dimension of the random vector represented by the distribution N.

MarkovKernels.entropy — Method

entropy(N::AbstractNormal)

Computes the entropy of the distribution N.

MarkovKernels.kldivergence — Method

kldivergence(N1::AbstractNormal, N2::AbstractNormal)

Computes the Kullback-Leibler divergence between N1 and N2.

MarkovKernels.logpdf — Method

logpdf(N::AbstractNormal,x)

Evaluates the logarithm of the probability density of N at x.

MarkovKernels.residual — Method

residual(N::AbstractNormal,x)

Given a realisation x, computes the whitened residual with respect to N.

Statistics.cov — Method

cov(N::AbstractNormal)

Returns the covariance matrix of the distribution N.

Statistics.mean — Method

mean(N::AbstractNormal)

Returns the mean vector of the distribution N.

Statistics.std — Method

std(N::AbstractNormal)

Returns a vector of marginal standard deviations of the distribution N.

Statistics.var — Method

var(N::AbstractNormal)

Returns a vector of marginal variances of the distribution N.

MarkovKernels.AbstractNormalKernel — Type

AbstractNormalKernel{T<:Number}

Abstract type for representing Normal kernels taking values in T.

MarkovKernels.NormalKernel — Type

NormalKernel

Standard parametrisation of Normal kernels.

MarkovKernels.condition — Method

condition(K::AbstractNormalKernel,x)

Returns a Normal distribution corresponding to K evaluated at x.

Statistics.cov — Method

cov(K::AbstractNormalKernel)

Returns the conditional covariance matrix of the Normal kernel K.

Statistics.mean — Method

mean(K::AbstractNormalKernel)

Returns the conditional mean function of the Normal kernel K.