Covariance Parameter

In MarkovKernels.jl a covariance matrix is assumed to be of the following type union.

const CovarianceParameter{T} = Union{HermOrSym{T},Factorization{T}}

Functions

MarkovKernels.lsqrt — Method

lsqrt(A::CovarianceParameter)

Computes a square matrix L such that A = L*L'. L need not be a Cholesky factor.

Missing docstring.

Missing docstring for stein(Σ, Φ::AbstractMatrix). Check Documenter's build log for details.

Missing docstring.

Missing docstring for stein(Σ, Φ::AbstractMatrix, Q). Check Documenter's build log for details.

Missing docstring.

Missing docstring for schur_reduce(Π, C::AbstractMatrix). Check Documenter's build log for details.

Missing docstring.

Missing docstring for schur_reduce(Π, C::AbstractMatrix, R). Check Documenter's build log for details.