DiracKernel

The Dirac kernel with conditional mean prameter $\mu$ is denotd by

\[k(y\mid x) = \delta(y - \mu(x)).\]

Types

MarkovKernels.AbstractDiracKernel — Type

AbstractDiracKernel

Abstract type for representing Dirac kernels.

MarkovKernels.DiracKernel — Type

DiracKernel

Type for representing Dirac kernels K(y,x) = δ(y - μ(x)).

MarkovKernels.IdentityKernel — Type

IdentityKernel

Struct for representing kernels that act like identity under marginalization.

Type aliases

const AffineDiracKernel{T} = DiracKernel{T,<:AbstractAffineMap}

Constructors

MarkovKernels.DiracKernel — Method

DiracKernel(Φ::AbstractMatrix, Σ)

Creates a DiracKernel with a linear conditional mean function given by

x ↦ Φ * x.

MarkovKernels.DiracKernel — Method

DiracKernel(Φ::AbstractMatrix, b::AbstractVector, Σ)

Creates a DiracKernel with an affine conditional mean function given by

x ↦ b + Φ * x.

MarkovKernels.DiracKernel — Method

DiracKernel(Φ::AbstractMatrix, b::AbstractVector, c::AbstractVector, Σ)

Creates a DiracKernel with an affine corrector conditional mean function given by

x ↦ b + Φ * (x - c).

Basics

Statistics.mean — Method

mean(K::AbstractDiracKernel)

Computes the conditonal mean function of the Dirac kernel K. That is, the output is callable.

Conditioning and sampling

MarkovKernels.condition — Method

condition(K::AbstractDiracKernel, x)

Returns a Dirac distribution corresponding to the Dirac kernel K evaluated at x.

Base.rand — Method

rand(::AbstractRNG, K::AbstractDiracKernel, x::AbstractVector)

Computes a random vector conditionally on x with respect the the Dirac kernel K using the random number generator RNG. Equivalent to mean(K)(x).

Base.rand — Method

rand(K::AbstractDiracKernel, x::AbstractVector)

Computes a random vector conditionally on x with respect the the Dirac kernel K using the random number generator Random.GLOBAL_RNG. Equivalent to mean(K)(x).