NormalKernel

The Normal kernel is denoted by

\[k(y\mid x) = \mathcal{N}(y ; \mu(x) , \Sigma(x) ).\]

As with the Normal distributions, the explicit expression on the kernel depends on whether it is real or complex valued.

Types

MarkovKernels.AbstractNormalKernel — Type

AbstractNormalKernel{T<:Number}

Abstract type for representing Normal kernels taking values in T.

MarkovKernels.NormalKernel — Type

NormalKernel

Standard mean vector / covariance matrix parametrisation of Normal kernels.

Type aliases

const AffineNormalKernel{T} = NormalKernel{T,<:AbstractAffineMap,<:CovarianceParameter}

Constructors

MarkovKernels.NormalKernel — Method

NormalKernel(F::AbstractAffineMap, Σ)

Creates a NormalKernel with conditional mean function F and conditional covariance function parameter Σ. Σ is assumed to be callable and be of compatible eltype with F.

MarkovKernels.NormalKernel — Method

NormalKernel(Φ::AbstractMatrix, Σ)

Creates a NormalKernel with a linear conditional mean function given by

x ↦ Φ * x,

and conditional covariance function parameter Σ. Σ is assumed to be callable and be of compatible eltype with Φ.

MarkovKernels.NormalKernel — Method

NormalKernel(Φ::AbstractMatrix, b::AbstractVector, Σ)

Creates a NormalKernel with an affine conditional mean function given by

x ↦ b + Φ * x,

and conditional covariance function parameter Σ. Σ is assumed to be callable and be of compatible eltype with Φ, b.

MarkovKernels.NormalKernel — Method

NormalKernel(Φ::AbstractMatrix, b::AbstractVector, c::AbstractVector, Σ)

Creates a NormalKernel with an affine corrector conditional mean function given by

x ↦ b + Φ * (x - c),

and conditional covariance function parameter Σ. Σ is assumed to be callable and be of compatible eltype with Φ, b, c.

MarkovKernels.NormalKernel — Method

NormalKernel(F::AbstractAffineMap, Σ::CovarianceParameter)

Creates a NormalKernel with conditional mean function F and a constant conditional covariance function parameterised by Σ.

MarkovKernels.NormalKernel — Method

NormalKernel(F::AbstractAffineMap, Σ::AbstractMatrix)

Creates a NormalKernel with conditional mean function F and a constant conditional covariance function Σ if Σ is Symmetric / Hermitian. Throws domain error otherwise.

MarkovKernels.NormalKernel — Method

NormalKernel{T}(K::AffineNormalKernel{U}) where {T,U}

Computes a Normal kernel of eltype T from the Normal kernel K if T and U are compatible. That is T and U must both be Real or both be Complex.

Basics

Statistics.mean — Method

mean(K::AbstractNormalKernel)

Computes the conditonal mean function of the Normal kernel K. That is, the output is callable.

Statistics.cov — Method

mean(K::AbstractNormalKernel)

Computes the conditonal covariance matrix function of the Normal kernel K. That is, the output is callable.

MarkovKernels.covp — Method

covp(K::AbstractNormalKernel)

Returns the internal representation of the conditonal covariance matrix of the Normal kernel K. For computing the actual conditional covariance matrix, use cov.

Conditioning and sampling

MarkovKernels.condition — Method

condition(K::AbstractNormalKernel, x)

Returns a Normal distribution corresponding to K evaluated at x.

Base.rand — Method

rand(RNG::AbstractRNG, K::AbstractNormalKernel, x::AbstractVector)

Computes a random vector conditionally on x with respect the the Normal kernel K using the random number generator RNG.

Base.rand — Method

rand(K::AbstractNormalKernel, x::AbstractVector)

Computes a random vector conditionally on x with respect the the Normal kernel K using the random number generator Random.GLOBAL_RNG.